Add svelte adapter static to build a static site
Trying this out to see the output of the HTML to see if it's usable for dropping into wordpress.

The HTML after formatting looks decent, but there's a lot of noise and overhead it seems like. Maybe on a fresh project without all the extras and layout would help out?